Two-Mark Determinantal Point Process Model

Implements a generic two-mark planar determinantal point process model. It is possible to set its parameters, fit the model to an observed planar point pattern or simulate a planar point pattern.

Active bindings

first_intensity

Setter and getter for the first intensity \(\rho_1\).

second_intensity

Setter and getter for the second intensity \(\rho_2\).

first_repulsion_rate

Setter and getter for the first repulsion rate \(\alpha_1\).

second_repulsion_rate

Setter and getter for the second repulsion rate \(\alpha_2\).

cross_repulsion_rate

Setter and getter for the cross repulsion rate \(\alpha_{12}\).

between_mark_correlation

Setter and getter for the correlation \(\tau\) between marks.

window_size

Setter and getter for the window size \(L\).

Methods

Public methods

TwoMarkDPP$new()

TwoMarkDPP$fit()

TwoMarkDPP$random()

TwoMarkDPP$print()

TwoMarkDPP$clone()

Method new()

The TwoMarkDPP constructor.

Usage

TwoMarkDPP$new(model = c("Gauss", "Bessel"))

Arguments

model

A string specifying the underlying parametric model to be used. Choices are "Gauss" or "Bessel". Defaults to "Gauss".

Returns

The TwoMarkDPP object invisibly.

Examples

mod <- TwoMarkDPP$new(model = "Gauss")

Method fit()

Estimates the parameters of the two-mark DPP from an observed planar point pattern.

Usage

TwoMarkDPP$fit(X, method = c("pcf", "mle"))

Arguments

X

An object of class spatstat.geom::ppp.

method

A string specifying the kind of estimator to use. Choices are either "pcf" or "mle". Defaults to "pcf".

Returns

The TwoMarkDPP object invisibly.

Examples

mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])

Method random()

Samples one or more planar point patterns according to the two-mark DPP.

Usage

TwoMarkDPP$random(n)

Arguments

n

An integer value specifying the sample size.

Returns

A base::list of objects of class spatstat.geom::ppp storing sampled planar point patterns.

Examples

mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])
mod$random(n = 1)

Method print()

Prints the TwoMarkDPP object.

Usage

TwoMarkDPP$print(...)

Arguments

...

Extra parameters to be passed on to next methods. Current unused.

Returns

NULL

Examples

mod <- TwoMarkDPP$new()
mod

Method clone()

The objects of this class are cloneable with this method.

Usage

TwoMarkDPP$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `TwoMarkDPP$new`
## ------------------------------------------------

mod <- TwoMarkDPP$new(model = "Gauss")

## ------------------------------------------------
## Method `TwoMarkDPP$fit`
## ------------------------------------------------

mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])

## ------------------------------------------------
## Method `TwoMarkDPP$random`
## ------------------------------------------------

mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])
mod$random(n = 1)
#> 2048
#> 1985
#> 1924
#> 1865
#> 1808
#> 1753
#> 1700
#> 1649
#> 1600
#> 1553
#> 1508
#> 1465
#> 1424
#> 1385
#> 1348
#> 1313
#> 1280
#> 1249
#> 1220
#> 1193
#> 1168
#> 1145
#> 1124
#> 1105
#> 1088
#> 1073
#> 1060
#> 1049
#> 1040
#> 1033
#> 1028
#> 1025
#> 1024
#> 1025
#> 1028
#> 1033
#> 1040
#> 1049
#> 1060
#> 1073
#> 1088
#> 1105
#> 1124
#> 1145
#> 1168
#> 1193
#> 1220
#> 1249
#> 1280
#> 1313
#> 1348
#> 1385
#> 1424
#> 1465
#> 1508
#> 1553
#> 1600
#> 1649
#> 1700
#> 1753
#> 1808
#> 1865
#> 1924
#> 1985
#> 2048
#> 1985
#> 1922
#> 1861
#> 1802
#> 1745
#> 1690
#> 1637
#> 1586
#> 1537
#> 1490
#> 1445
#> 1402
#> 1361
#> 1322
#> 1285
#> 1250
#> 1217
#> 1186
#> 1157
#> 1130
#> 1105
#> 1082
#> 1061
#> 1042
#> 1025
#> 1010
#> 997
#> 986
#> 977
#> 970
#> 965
#> 962
#> 961
#> 962
#> 965
#> 970
#> 977
#> 986
#> 997
#> 1010
#> 1025
#> 1042
#> 1061
#> 1082
#> 1105
#> 1130
#> 1157
#> 1186
#> 1217
#> 1250
#> 1285
#> 1322
#> 1361
#> 1402
#> 1445
#> 1490
#> 1537
#> 1586
#> 1637
#> 1690
#> 1745
#> 1802
#> 1861
#> 1922
#> 1985
#> 1924
#> 1861
#> 1800
#> 1741
#> 1684
#> 1629
#> 1576
#> 1525
#> 1476
#> 1429
#> 1384
#> 1341
#> 1300
#> 1261
#> 1224
#> 1189
#> 1156
#> 1125
#> 1096
#> 1069
#> 1044
#> 1021
#> 1000
#> 981
#> 964
#> 949
#> 936
#> 925
#> 916
#> 909
#> 904
#> 901
#> 900
#> 901
#> 904
#> 909
#> 916
#> 925
#> 936
#> 949
#> 964
#> 981
#> 1000
#> 1021
#> 1044
#> 1069
#> 1096
#> 1125
#> 1156
#> 1189
#> 1224
#> 1261
#> 1300
#> 1341
#> 1384
#> 1429
#> 1476
#> 1525
#> 1576
#> 1629
#> 1684
#> 1741
#> 1800
#> 1861
#> 1924
#> 1865
#> 1802
#> 1741
#> 1682
#> 1625
#> 1570
#> 1517
#> 1466
#> 1417
#> 1370
#> 1325
#> 1282
#> 1241
#> 1202
#> 1165
#> 1130
#> 1097
#> 1066
#> 1037
#> 1010
#> 985
#> 962
#> 941
#> 922
#> 905
#> 890
#> 877
#> 866
#> 857
#> 850
#> 845
#> 842
#> 841
#> 842
#> 845
#> 850
#> 857
#> 866
#> 877
#> 890
#> 905
#> 922
#> 941
#> 962
#> 985
#> 1010
#> 1037
#> 1066
#> 1097
#> 1130
#> 1165
#> 1202
#> 1241
#> 1282
#> 1325
#> 1370
#> 1417
#> 1466
#> 1517
#> 1570
#> 1625
#> 1682
#> 1741
#> 1802
#> 1865
#> 1808
#> 1745
#> 1684
#> 1625
#> 1568
#> 1513
#> 1460
#> 1409
#> 1360
#> 1313
#> 1268
#> 1225
#> 1184
#> 1145
#> 1108
#> 1073
#> 1040
#> 1009
#> 980
#> 953
#> 928
#> 905
#> 884
#> 865
#> 848
#> 833
#> 820
#> 809
#> 800
#> 793
#> 788
#> 785
#> 784
#> 785
#> 788
#> 793
#> 800
#> 809
#> 820
#> 833
#> 848
#> 865
#> 884
#> 905
#> 928
#> 953
#> 980
#> 1009
#> 1040
#> 1073
#> 1108
#> 1145
#> 1184
#> 1225
#> 1268
#> 1313
#> 1360
#> 1409
#> 1460
#> 1513
#> 1568
#> 1625
#> 1684
#> 1745
#> 1808
#> 1753
#> 1690
#> 1629
#> 1570
#> 1513
#> 1458
#> 1405
#> 1354
#> 1305
#> 1258
#> 1213
#> 1170
#> 1129
#> 1090
#> 1053
#> 1018
#> 985
#> 954
#> 925
#> 898
#> 873
#> 850
#> 829
#> 810
#> 793
#> 778
#> 765
#> 754
#> 745
#> 738
#> 733
#> 730
#> 729
#> 730
#> 733
#> 738
#> 745
#> 754
#> 765
#> 778
#> 793
#> 810
#> 829
#> 850
#> 873
#> 898
#> 925
#> 954
#> 985
#> 1018
#> 1053
#> 1090
#> 1129
#> 1170
#> 1213
#> 1258
#> 1305
#> 1354
#> 1405
#> 1458
#> 1513
#> 1570
#> 1629
#> 1690
#> 1753
#> 1700
#> 1637
#> 1576
#> 1517
#> 1460
#> 1405
#> 1352
#> 1301
#> 1252
#> 1205
#> 1160
#> 1117
#> 1076
#> 1037
#> 1000
#> 965
#> 932
#> 901
#> 872
#> 845
#> 820
#> 797
#> 776
#> 757
#> 740
#> 725
#> 712
#> 701
#> 692
#> 685
#> 680
#> 677
#> 676
#> 677
#> 680
#> 685
#> 692
#> 701
#> 712
#> 725
#> 740
#> 757
#> 776
#> 797
#> 820
#> 845
#> 872
#> 901
#> 932
#> 965
#> 1000
#> 1037
#> 1076
#> 1117
#> 1160
#> 1205
#> 1252
#> 1301
#> 1352
#> 1405
#> 1460
#> 1517
#> 1576
#> 1637
#> 1700
#> 1649
#> 1586
#> 1525
#> 1466
#> 1409
#> 1354
#> 1301
#> 1250
#> 1201
#> 1154
#> 1109
#> 1066
#> 1025
#> 986
#> 949
#> 914
#> 881
#> 850
#> 821
#> 794
#> 769
#> 746
#> 725
#> 706
#> 689
#> 674
#> 661
#> 650
#> 641
#> 634
#> 629
#> 626
#> 625
#> 626
#> 629
#> 634
#> 641
#> 650
#> 661
#> 674
#> 689
#> 706
#> 725
#> 746
#> 769
#> 794
#> 821
#> 850
#> 881
#> 914
#> 949
#> 986
#> 1025
#> 1066
#> 1109
#> 1154
#> 1201
#> 1250
#> 1301
#> 1354
#> 1409
#> 1466
#> 1525
#> 1586
#> 1649
#> 1600
#> 1537
#> 1476
#> 1417
#> 1360
#> 1305
#> 1252
#> 1201
#> 1152
#> 1105
#> 1060
#> 1017
#> 976
#> 937
#> 900
#> 865
#> 832
#> 801
#> 772
#> 745
#> 720
#> 697
#> 676
#> 657
#> 640
#> 625
#> 612
#> 601
#> 592
#> 585
#> 580
#> 577
#> 576
#> 577
#> 580
#> 585
#> 592
#> 601
#> 612
#> 625
#> 640
#> 657
#> 676
#> 697
#> 720
#> 745
#> 772
#> 801
#> 832
#> 865
#> 900
#> 937
#> 976
#> 1017
#> 1060
#> 1105
#> 1152
#> 1201
#> 1252
#> 1305
#> 1360
#> 1417
#> 1476
#> 1537
#> 1600
#> 1553
#> 1490
#> 1429
#> 1370
#> 1313
#> 1258
#> 1205
#> 1154
#> 1105
#> 1058
#> 1013
#> 970
#> 929
#> 890
#> 853
#> 818
#> 785
#> 754
#> 725
#> 698
#> 673
#> 650
#> 629
#> 610
#> 593
#> 578
#> 565
#> 554
#> 545
#> 538
#> 533
#> 530
#> 529
#> 530
#> 533
#> 538
#> 545
#> 554
#> 565
#> 578
#> 593
#> 610
#> 629
#> 650
#> 673
#> 698
#> 725
#> 754
#> 785
#> 818
#> 853
#> 890
#> 929
#> 970
#> 1013
#> 1058
#> 1105
#> 1154
#> 1205
#> 1258
#> 1313
#> 1370
#> 1429
#> 1490
#> 1553
#> 1508
#> 1445
#> 1384
#> 1325
#> 1268
#> 1213
#> 1160
#> 1109
#> 1060
#> 1013
#> 968
#> 925
#> 884
#> 845
#> 808
#> 773
#> 740
#> 709
#> 680
#> 653
#> 628
#> 605
#> 584
#> 565
#> 548
#> 533
#> 520
#> 509
#> 500
#> 493
#> 488
#> 485
#> 484
#> 485
#> 488
#> 493
#> 500
#> 509
#> 520
#> 533
#> 548
#> 565
#> 584
#> 605
#> 628
#> 653
#> 680
#> 709
#> 740
#> 773
#> 808
#> 845
#> 884
#> 925
#> 968
#> 1013
#> 1060
#> 1109
#> 1160
#> 1213
#> 1268
#> 1325
#> 1384
#> 1445
#> 1508
#> 1465
#> 1402
#> 1341
#> 1282
#> 1225
#> 1170
#> 1117
#> 1066
#> 1017
#> 970
#> 925
#> 882
#> 841
#> 802
#> 765
#> 730
#> 697
#> 666
#> 637
#> 610
#> 585
#> 562
#> 541
#> 522
#> 505
#> 490
#> 477
#> 466
#> 457
#> 450
#> 445
#> 442
#> 441
#> 442
#> 445
#> 450
#> 457
#> 466
#> 477
#> 490
#> 505
#> 522
#> 541
#> 562
#> 585
#> 610
#> 637
#> 666
#> 697
#> 730
#> 765
#> 802
#> 841
#> 882
#> 925
#> 970
#> 1017
#> 1066
#> 1117
#> 1170
#> 1225
#> 1282
#> 1341
#> 1402
#> 1465
#> 1424
#> 1361
#> 1300
#> 1241
#> 1184
#> 1129
#> 1076
#> 1025
#> 976
#> 929
#> 884
#> 841
#> 800
#> 761
#> 724
#> 689
#> 656
#> 625
#> 596
#> 569
#> 544
#> 521
#> 500
#> 481
#> 464
#> 449
#> 436
#> 425
#> 416
#> 409
#> 404
#> 401
#> 400
#> 401
#> 404
#> 409
#> 416
#> 425
#> 436
#> 449
#> 464
#> 481
#> 500
#> 521
#> 544
#> 569
#> 596
#> 625
#> 656
#> 689
#> 724
#> 761
#> 800
#> 841
#> 884
#> 929
#> 976
#> 1025
#> 1076
#> 1129
#> 1184
#> 1241
#> 1300
#> 1361
#> 1424
#> 1385
#> 1322
#> 1261
#> 1202
#> 1145
#> 1090
#> 1037
#> 986
#> 937
#> 890
#> 845
#> 802
#> 761
#> 722
#> 685
#> 650
#> 617
#> 586
#> 557
#> 530
#> 505
#> 482
#> 461
#> 442
#> 425
#> 410
#> 397
#> 386
#> 377
#> 370
#> 365
#> 362
#> 361
#> 362
#> 365
#> 370
#> 377
#> 386
#> 397
#> 410
#> 425
#> 442
#> 461
#> 482
#> 505
#> 530
#> 557
#> 586
#> 617
#> 650
#> 685
#> 722
#> 761
#> 802
#> 845
#> 890
#> 937
#> 986
#> 1037
#> 1090
#> 1145
#> 1202
#> 1261
#> 1322
#> 1385
#> 1348
#> 1285
#> 1224
#> 1165
#> 1108
#> 1053
#> 1000
#> 949
#> 900
#> 853
#> 808
#> 765
#> 724
#> 685
#> 648
#> 613
#> 580
#> 549
#> 520
#> 493
#> 468
#> 445
#> 424
#> 405
#> 388
#> 373
#> 360
#> 349
#> 340
#> 333
#> 328
#> 325
#> 324
#> 325
#> 328
#> 333
#> 340
#> 349
#> 360
#> 373
#> 388
#> 405
#> 424
#> 445
#> 468
#> 493
#> 520
#> 549
#> 580
#> 613
#> 648
#> 685
#> 724
#> 765
#> 808
#> 853
#> 900
#> 949
#> 1000
#> 1053
#> 1108
#> 1165
#> 1224
#> 1285
#> 1348
#> 1313
#> 1250
#> 1189
#> 1130
#> 1073
#> 1018
#> 965
#> 914
#> 865
#> 818
#> 773
#> 730
#> 689
#> 650
#> 613
#> 578
#> 545
#> 514
#> 485
#> 458
#> 433
#> 410
#> 389
#> 370
#> 353
#> 338
#> 325
#> 314
#> 305
#> 298
#> 293
#> 290
#> 289
#> 290
#> 293
#> 298
#> 305
#> 314
#> 325
#> 338
#> 353
#> 370
#> 389
#> 410
#> 433
#> 458
#> 485
#> 514
#> 545
#> 578
#> 613
#> 650
#> 689
#> 730
#> 773
#> 818
#> 865
#> 914
#> 965
#> 1018
#> 1073
#> 1130
#> 1189
#> 1250
#> 1313
#> 1280
#> 1217
#> 1156
#> 1097
#> 1040
#> 985
#> 932
#> 881
#> 832
#> 785
#> 740
#> 697
#> 656
#> 617
#> 580
#> 545
#> 512
#> 481
#> 452
#> 425
#> 400
#> 377
#> 356
#> 337
#> 320
#> 305
#> 292
#> 281
#> 272
#> 265
#> 260
#> 257
#> 256
#> 257
#> 260
#> 265
#> 272
#> 281
#> 292
#> 305
#> 320
#> 337
#> 356
#> 377
#> 400
#> 425
#> 452
#> 481
#> 512
#> 545
#> 580
#> 617
#> 656
#> 697
#> 740
#> 785
#> 832
#> 881
#> 932
#> 985
#> 1040
#> 1097
#> 1156
#> 1217
#> 1280
#> 1249
#> 1186
#> 1125
#> 1066
#> 1009
#> 954
#> 901
#> 850
#> 801
#> 754
#> 709
#> 666
#> 625
#> 586
#> 549
#> 514
#> 481
#> 450
#> 421
#> 394
#> 369
#> 346
#> 325
#> 306
#> 289
#> 274
#> 261
#> 250
#> 241
#> 234
#> 229
#> 226
#> 225
#> 226
#> 229
#> 234
#> 241
#> 250
#> 261
#> 274
#> 289
#> 306
#> 325
#> 346
#> 369
#> 394
#> 421
#> 450
#> 481
#> 514
#> 549
#> 586
#> 625
#> 666
#> 709
#> 754
#> 801
#> 850
#> 901
#> 954
#> 1009
#> 1066
#> 1125
#> 1186
#> 1249
#> 1220
#> 1157
#> 1096
#> 1037
#> 980
#> 925
#> 872
#> 821
#> 772
#> 725
#> 680
#> 637
#> 596
#> 557
#> 520
#> 485
#> 452
#> 421
#> 392
#> 365
#> 340
#> 317
#> 296
#> 277
#> 260
#> 245
#> 232
#> 221
#> 212
#> 205
#> 200
#> 197
#> 196
#> 197
#> 200
#> 205
#> 212
#> 221
#> 232
#> 245
#> 260
#> 277
#> 296
#> 317
#> 340
#> 365
#> 392
#> 421
#> 452
#> 485
#> 520
#> 557
#> 596
#> 637
#> 680
#> 725
#> 772
#> 821
#> 872
#> 925
#> 980
#> 1037
#> 1096
#> 1157
#> 1220
#> 1193
#> 1130
#> 1069
#> 1010
#> 953
#> 898
#> 845
#> 794
#> 745
#> 698
#> 653
#> 610
#> 569
#> 530
#> 493
#> 458
#> 425
#> 394
#> 365
#> 338
#> 313
#> 290
#> 269
#> 250
#> 233
#> 218
#> 205
#> 194
#> 185
#> 178
#> 173
#> 170
#> 169
#> 170
#> 173
#> 178
#> 185
#> 194
#> 205
#> 218
#> 233
#> 250
#> 269
#> 290
#> 313
#> 338
#> 365
#> 394
#> 425
#> 458
#> 493
#> 530
#> 569
#> 610
#> 653
#> 698
#> 745
#> 794
#> 845
#> 898
#> 953
#> 1010
#> 1069
#> 1130
#> 1193
#> 1168
#> 1105
#> 1044
#> 985
#> 928
#> 873
#> 820
#> 769
#> 720
#> 673
#> 628
#> 585
#> 544
#> 505
#> 468
#> 433
#> 400
#> 369
#> 340
#> 313
#> 288
#> 265
#> 244
#> 225
#> 208
#> 193
#> 180
#> 169
#> 160
#> 153
#> 148
#> 145
#> 144
#> 145
#> 148
#> 153
#> 160
#> 169
#> 180
#> 193
#> 208
#> 225
#> 244
#> 265
#> 288
#> 313
#> 340
#> 369
#> 400
#> 433
#> 468
#> 505
#> 544
#> 585
#> 628
#> 673
#> 720
#> 769
#> 820
#> 873
#> 928
#> 985
#> 1044
#> 1105
#> 1168
#> 1145
#> 1082
#> 1021
#> 962
#> 905
#> 850
#> 797
#> 746
#> 697
#> 650
#> 605
#> 562
#> 521
#> 482
#> 445
#> 410
#> 377
#> 346
#> 317
#> 290
#> 265
#> 242
#> 221
#> 202
#> 185
#> 170
#> 157
#> 146
#> 137
#> 130
#> 125
#> 122
#> 121
#> 122
#> 125
#> 130
#> 137
#> 146
#> 157
#> 170
#> 185
#> 202
#> 221
#> 242
#> 265
#> 290
#> 317
#> 346
#> 377
#> 410
#> 445
#> 482
#> 521
#> 562
#> 605
#> 650
#> 697
#> 746
#> 797
#> 850
#> 905
#> 962
#> 1021
#> 1082
#> 1145
#> 1124
#> 1061
#> 1000
#> 941
#> 884
#> 829
#> 776
#> 725
#> 676
#> 629
#> 584
#> 541
#> 500
#> 461
#> 424
#> 389
#> 356
#> 325
#> 296
#> 269
#> 244
#> 221
#> 200
#> 181
#> 164
#> 149
#> 136
#> 125
#> 116
#> 109
#> 104
#> 101
#> 100
#> 101
#> 104
#> 109
#> 116
#> 125
#> 136
#> 149
#> 164
#> 181
#> 200
#> 221
#> 244
#> 269
#> 296
#> 325
#> 356
#> 389
#> 424
#> 461
#> 500
#> 541
#> 584
#> 629
#> 676
#> 725
#> 776
#> 829
#> 884
#> 941
#> 1000
#> 1061
#> 1124
#> 1105
#> 1042
#> 981
#> 922
#> 865
#> 810
#> 757
#> 706
#> 657
#> 610
#> 565
#> 522
#> 481
#> 442
#> 405
#> 370
#> 337
#> 306
#> 277
#> 250
#> 225
#> 202
#> 181
#> 162
#> 145
#> 130
#> 117
#> 106
#> 97
#> 90
#> 85
#> 82
#> 81
#> 82
#> 85
#> 90
#> 97
#> 106
#> 117
#> 130
#> 145
#> 162
#> 181
#> 202
#> 225
#> 250
#> 277
#> 306
#> 337
#> 370
#> 405
#> 442
#> 481
#> 522
#> 565
#> 610
#> 657
#> 706
#> 757
#> 810
#> 865
#> 922
#> 981
#> 1042
#> 1105
#> 1088
#> 1025
#> 964
#> 905
#> 848
#> 793
#> 740
#> 689
#> 640
#> 593
#> 548
#> 505
#> 464
#> 425
#> 388
#> 353
#> 320
#> 289
#> 260
#> 233
#> 208
#> 185
#> 164
#> 145
#> 128
#> 113
#> 100
#> 89
#> 80
#> 73
#> 68
#> 65
#> 64
#> 65
#> 68
#> 73
#> 80
#> 89
#> 100
#> 113
#> 128
#> 145
#> 164
#> 185
#> 208
#> 233
#> 260
#> 289
#> 320
#> 353
#> 388
#> 425
#> 464
#> 505
#> 548
#> 593
#> 640
#> 689
#> 740
#> 793
#> 848
#> 905
#> 964
#> 1025
#> 1088
#> 1073
#> 1010
#> 949
#> 890
#> 833
#> 778
#> 725
#> 674
#> 625
#> 578
#> 533
#> 490
#> 449
#> 410
#> 373
#> 338
#> 305
#> 274
#> 245
#> 218
#> 193
#> 170
#> 149
#> 130
#> 113
#> 98
#> 85
#> 74
#> 65
#> 58
#> 53
#> 50
#> 49
#> 50
#> 53
#> 58
#> 65
#> 74
#> 85
#> 98
#> 113
#> 130
#> 149
#> 170
#> 193
#> 218
#> 245
#> 274
#> 305
#> 338
#> 373
#> 410
#> 449
#> 490
#> 533
#> 578
#> 625
#> 674
#> 725
#> 778
#> 833
#> 890
#> 949
#> 1010
#> 1073
#> 1060
#> 997
#> 936
#> 877
#> 820
#> 765
#> 712
#> 661
#> 612
#> 565
#> 520
#> 477
#> 436
#> 397
#> 360
#> 325
#> 292
#> 261
#> 232
#> 205
#> 180
#> 157
#> 136
#> 117
#> 100
#> 85
#> 72
#> 61
#> 52
#> 45
#> 40
#> 37
#> 36
#> 37
#> 40
#> 45
#> 52
#> 61
#> 72
#> 85
#> 100
#> 117
#> 136
#> 157
#> 180
#> 205
#> 232
#> 261
#> 292
#> 325
#> 360
#> 397
#> 436
#> 477
#> 520
#> 565
#> 612
#> 661
#> 712
#> 765
#> 820
#> 877
#> 936
#> 997
#> 1060
#> 1049
#> 986
#> 925
#> 866
#> 809
#> 754
#> 701
#> 650
#> 601
#> 554
#> 509
#> 466
#> 425
#> 386
#> 349
#> 314
#> 281
#> 250
#> 221
#> 194
#> 169
#> 146
#> 125
#> 106
#> 89
#> 74
#> 61
#> 50
#> 41
#> 34
#> 29
#> 26
#> 25
#> 26
#> 29
#> 34
#> 41
#> 50
#> 61
#> 74
#> 89
#> 106
#> 125
#> 146
#> 169
#> 194
#> 221
#> 250
#> 281
#> 314
#> 349
#> 386
#> 425
#> 466
#> 509
#> 554
#> 601
#> 650
#> 701
#> 754
#> 809
#> 866
#> 925
#> 986
#> 1049
#> 1040
#> 977
#> 916
#> 857
#> 800
#> 745
#> 692
#> 641
#> 592
#> 545
#> 500
#> 457
#> 416
#> 377
#> 340
#> 305
#> 272
#> 241
#> 212
#> 185
#> 160
#> 137
#> 116
#> 97
#> 80
#> 65
#> 52
#> 41
#> 32
#> 25
#> 20
#> 17
#> 16
#> 17
#> 20
#> 25
#> 32
#> 41
#> 52
#> 65
#> 80
#> 97
#> 116
#> 137
#> 160
#> 185
#> 212
#> 241
#> 272
#> 305
#> 340
#> 377
#> 416
#> 457
#> 500
#> 545
#> 592
#> 641
#> 692
#> 745
#> 800
#> 857
#> 916
#> 977
#> 1040
#> 1033
#> 970
#> 909
#> 850
#> 793
#> 738
#> 685
#> 634
#> 585
#> 538
#> 493
#> 450
#> 409
#> 370
#> 333
#> 298
#> 265
#> 234
#> 205
#> 178
#> 153
#> 130
#> 109
#> 90
#> 73
#> 58
#> 45
#> 34
#> 25
#> 18
#> 13
#> 10
#> 9
#> 10
#> 13
#> 18
#> 25
#> 34
#> 45
#> 58
#> 73
#> 90
#> 109
#> 130
#> 153
#> 178
#> 205
#> 234
#> 265
#> 298
#> 333
#> 370
#> 409
#> 450
#> 493
#> 538
#> 585
#> 634
#> 685
#> 738
#> 793
#> 850
#> 909
#> 970
#> 1033
#> 1028
#> 965
#> 904
#> 845
#> 788
#> 733
#> 680
#> 629
#> 580
#> 533
#> 488
#> 445
#> 404
#> 365
#> 328
#> 293
#> 260
#> 229
#> 200
#> 173
#> 148
#> 125
#> 104
#> 85
#> 68
#> 53
#> 40
#> 29
#> 20
#> 13
#> 8
#> 5
#> 4
#> 5
#> 8
#> 13
#> 20
#> 29
#> 40
#> 53
#> 68
#> 85
#> 104
#> 125
#> 148
#> 173
#> 200
#> 229
#> 260
#> 293
#> 328
#> 365
#> 404
#> 445
#> 488
#> 533
#> 580
#> 629
#> 680
#> 733
#> 788
#> 845
#> 904
#> 965
#> 1028
#> 1025
#> 962
#> 901
#> 842
#> 785
#> 730
#> 677
#> 626
#> 577
#> 530
#> 485
#> 442
#> 401
#> 362
#> 325
#> 290
#> 257
#> 226
#> 197
#> 170
#> 145
#> 122
#> 101
#> 82
#> 65
#> 50
#> 37
#> 26
#> 17
#> 10
#> 5
#> 2
#> 1
#> 2
#> 5
#> 10
#> 17
#> 26
#> 37
#> 50
#> 65
#> 82
#> 101
#> 122
#> 145
#> 170
#> 197
#> 226
#> 257
#> 290
#> 325
#> 362
#> 401
#> 442
#> 485
#> 530
#> 577
#> 626
#> 677
#> 730
#> 785
#> 842
#> 901
#> 962
#> 1025
#> 1024
#> 961
#> 900
#> 841
#> 784
#> 729
#> 676
#> 625
#> 576
#> 529
#> 484
#> 441
#> 400
#> 361
#> 324
#> 289
#> 256
#> 225
#> 196
#> 169
#> 144
#> 121
#> 100
#> 81
#> 64
#> 49
#> 36
#> 25
#> 16
#> 9
#> 4
#> 1
#> 0
#> 1
#> 4
#> 9
#> 16
#> 25
#> 36
#> 49
#> 64
#> 81
#> 100
#> 121
#> 144
#> 169
#> 196
#> 225
#> 256
#> 289
#> 324
#> 361
#> 400
#> 441
#> 484
#> 529
#> 576
#> 625
#> 676
#> 729
#> 784
#> 841
#> 900
#> 961
#> 1024
#> 1025
#> 962
#> 901
#> 842
#> 785
#> 730
#> 677
#> 626
#> 577
#> 530
#> 485
#> 442
#> 401
#> 362
#> 325
#> 290
#> 257
#> 226
#> 197
#> 170
#> 145
#> 122
#> 101
#> 82
#> 65
#> 50
#> 37
#> 26
#> 17
#> 10
#> 5
#> 2
#> 1
#> 2
#> 5
#> 10
#> 17
#> 26
#> 37
#> 50
#> 65
#> 82
#> 101
#> 122
#> 145
#> 170
#> 197
#> 226
#> 257
#> 290
#> 325
#> 362
#> 401
#> 442
#> 485
#> 530
#> 577
#> 626
#> 677
#> 730
#> 785
#> 842
#> 901
#> 962
#> 1025
#> 1028
#> 965
#> 904
#> 845
#> 788
#> 733
#> 680
#> 629
#> 580
#> 533
#> 488
#> 445
#> 404
#> 365
#> 328
#> 293
#> 260
#> 229
#> 200
#> 173
#> 148
#> 125
#> 104
#> 85
#> 68
#> 53
#> 40
#> 29
#> 20
#> 13
#> 8
#> 5
#> 4
#> 5
#> 8
#> 13
#> 20
#> 29
#> 40
#> 53
#> 68
#> 85
#> 104
#> 125
#> 148
#> 173
#> 200
#> 229
#> 260
#> 293
#> 328
#> 365
#> 404
#> 445
#> 488
#> 533
#> 580
#> 629
#> 680
#> 733
#> 788
#> 845
#> 904
#> 965
#> 1028
#> 1033
#> 970
#> 909
#> 850
#> 793
#> 738
#> 685
#> 634
#> 585
#> 538
#> 493
#> 450
#> 409
#> 370
#> 333
#> 298
#> 265
#> 234
#> 205
#> 178
#> 153
#> 130
#> 109
#> 90
#> 73
#> 58
#> 45
#> 34
#> 25
#> 18
#> 13
#> 10
#> 9
#> 10
#> 13
#> 18
#> 25
#> 34
#> 45
#> 58
#> 73
#> 90
#> 109
#> 130
#> 153
#> 178
#> 205
#> 234
#> 265
#> 298
#> 333
#> 370
#> 409
#> 450
#> 493
#> 538
#> 585
#> 634
#> 685
#> 738
#> 793
#> 850
#> 909
#> 970
#> 1033
#> 1040
#> 977
#> 916
#> 857
#> 800
#> 745
#> 692
#> 641
#> 592
#> 545
#> 500
#> 457
#> 416
#> 377
#> 340
#> 305
#> 272
#> 241
#> 212
#> 185
#> 160
#> 137
#> 116
#> 97
#> 80
#> 65
#> 52
#> 41
#> 32
#> 25
#> 20
#> 17
#> 16
#> 17
#> 20
#> 25
#> 32
#> 41
#> 52
#> 65
#> 80
#> 97
#> 116
#> 137
#> 160
#> 185
#> 212
#> 241
#> 272
#> 305
#> 340
#> 377
#> 416
#> 457
#> 500
#> 545
#> 592
#> 641
#> 692
#> 745
#> 800
#> 857
#> 916
#> 977
#> 1040
#> 1049
#> 986
#> 925
#> 866
#> 809
#> 754
#> 701
#> 650
#> 601
#> 554
#> 509
#> 466
#> 425
#> 386
#> 349
#> 314
#> 281
#> 250
#> 221
#> 194
#> 169
#> 146
#> 125
#> 106
#> 89
#> 74
#> 61
#> 50
#> 41
#> 34
#> 29
#> 26
#> 25
#> 26
#> 29
#> 34
#> 41
#> 50
#> 61
#> 74
#> 89
#> 106
#> 125
#> 146
#> 169
#> 194
#> 221
#> 250
#> 281
#> 314
#> 349
#> 386
#> 425
#> 466
#> 509
#> 554
#> 601
#> 650
#> 701
#> 754
#> 809
#> 866
#> 925
#> 986
#> 1049
#> 1060
#> 997
#> 936
#> 877
#> 820
#> 765
#> 712
#> 661
#> 612
#> 565
#> 520
#> 477
#> 436
#> 397
#> 360
#> 325
#> 292
#> 261
#> 232
#> 205
#> 180
#> 157
#> 136
#> 117
#> 100
#> 85
#> 72
#> 61
#> 52
#> 45
#> 40
#> 37
#> 36
#> 37
#> 40
#> 45
#> 52
#> 61
#> 72
#> 85
#> 100
#> 117
#> 136
#> 157
#> 180
#> 205
#> 232
#> 261
#> 292
#> 325
#> 360
#> 397
#> 436
#> 477
#> 520
#> 565
#> 612
#> 661
#> 712
#> 765
#> 820
#> 877
#> 936
#> 997
#> 1060
#> 1073
#> 1010
#> 949
#> 890
#> 833
#> 778
#> 725
#> 674
#> 625
#> 578
#> 533
#> 490
#> 449
#> 410
#> 373
#> 338
#> 305
#> 274
#> 245
#> 218
#> 193
#> 170
#> 149
#> 130
#> 113
#> 98
#> 85
#> 74
#> 65
#> 58
#> 53
#> 50
#> 49
#> 50
#> 53
#> 58
#> 65
#> 74
#> 85
#> 98
#> 113
#> 130
#> 149
#> 170
#> 193
#> 218
#> 245
#> 274
#> 305
#> 338
#> 373
#> 410
#> 449
#> 490
#> 533
#> 578
#> 625
#> 674
#> 725
#> 778
#> 833
#> 890
#> 949
#> 1010
#> 1073
#> 1088
#> 1025
#> 964
#> 905
#> 848
#> 793
#> 740
#> 689
#> 640
#> 593
#> 548
#> 505
#> 464
#> 425
#> 388
#> 353
#> 320
#> 289
#> 260
#> 233
#> 208
#> 185
#> 164
#> 145
#> 128
#> 113
#> 100
#> 89
#> 80
#> 73
#> 68
#> 65
#> 64
#> 65
#> 68
#> 73
#> 80
#> 89
#> 100
#> 113
#> 128
#> 145
#> 164
#> 185
#> 208
#> 233
#> 260
#> 289
#> 320
#> 353
#> 388
#> 425
#> 464
#> 505
#> 548
#> 593
#> 640
#> 689
#> 740
#> 793
#> 848
#> 905
#> 964
#> 1025
#> 1088
#> 1105
#> 1042
#> 981
#> 922
#> 865
#> 810
#> 757
#> 706
#> 657
#> 610
#> 565
#> 522
#> 481
#> 442
#> 405
#> 370
#> 337
#> 306
#> 277
#> 250
#> 225
#> 202
#> 181
#> 162
#> 145
#> 130
#> 117
#> 106
#> 97
#> 90
#> 85
#> 82
#> 81
#> 82
#> 85
#> 90
#> 97
#> 106
#> 117
#> 130
#> 145
#> 162
#> 181
#> 202
#> 225
#> 250
#> 277
#> 306
#> 337
#> 370
#> 405
#> 442
#> 481
#> 522
#> 565
#> 610
#> 657
#> 706
#> 757
#> 810
#> 865
#> 922
#> 981
#> 1042
#> 1105
#> 1124
#> 1061
#> 1000
#> 941
#> 884
#> 829
#> 776
#> 725
#> 676
#> 629
#> 584
#> 541
#> 500
#> 461
#> 424
#> 389
#> 356
#> 325
#> 296
#> 269
#> 244
#> 221
#> 200
#> 181
#> 164
#> 149
#> 136
#> 125
#> 116
#> 109
#> 104
#> 101
#> 100
#> 101
#> 104
#> 109
#> 116
#> 125
#> 136
#> 149
#> 164
#> 181
#> 200
#> 221
#> 244
#> 269
#> 296
#> 325
#> 356
#> 389
#> 424
#> 461
#> 500
#> 541
#> 584
#> 629
#> 676
#> 725
#> 776
#> 829
#> 884
#> 941
#> 1000
#> 1061
#> 1124
#> 1145
#> 1082
#> 1021
#> 962
#> 905
#> 850
#> 797
#> 746
#> 697
#> 650
#> 605
#> 562
#> 521
#> 482
#> 445
#> 410
#> 377
#> 346
#> 317
#> 290
#> 265
#> 242
#> 221
#> 202
#> 185
#> 170
#> 157
#> 146
#> 137
#> 130
#> 125
#> 122
#> 121
#> 122
#> 125
#> 130
#> 137
#> 146
#> 157
#> 170
#> 185
#> 202
#> 221
#> 242
#> 265
#> 290
#> 317
#> 346
#> 377
#> 410
#> 445
#> 482
#> 521
#> 562
#> 605
#> 650
#> 697
#> 746
#> 797
#> 850
#> 905
#> 962
#> 1021
#> 1082
#> 1145
#> 1168
#> 1105
#> 1044
#> 985
#> 928
#> 873
#> 820
#> 769
#> 720
#> 673
#> 628
#> 585
#> 544
#> 505
#> 468
#> 433
#> 400
#> 369
#> 340
#> 313
#> 288
#> 265
#> 244
#> 225
#> 208
#> 193
#> 180
#> 169
#> 160
#> 153
#> 148
#> 145
#> 144
#> 145
#> 148
#> 153
#> 160
#> 169
#> 180
#> 193
#> 208
#> 225
#> 244
#> 265
#> 288
#> 313
#> 340
#> 369
#> 400
#> 433
#> 468
#> 505
#> 544
#> 585
#> 628
#> 673
#> 720
#> 769
#> 820
#> 873
#> 928
#> 985
#> 1044
#> 1105
#> 1168
#> 1193
#> 1130
#> 1069
#> 1010
#> 953
#> 898
#> 845
#> 794
#> 745
#> 698
#> 653
#> 610
#> 569
#> 530
#> 493
#> 458
#> 425
#> 394
#> 365
#> 338
#> 313
#> 290
#> 269
#> 250
#> 233
#> 218
#> 205
#> 194
#> 185
#> 178
#> 173
#> 170
#> 169
#> 170
#> 173
#> 178
#> 185
#> 194
#> 205
#> 218
#> 233
#> 250
#> 269
#> 290
#> 313
#> 338
#> 365
#> 394
#> 425
#> 458
#> 493
#> 530
#> 569
#> 610
#> 653
#> 698
#> 745
#> 794
#> 845
#> 898
#> 953
#> 1010
#> 1069
#> 1130
#> 1193
#> 1220
#> 1157
#> 1096
#> 1037
#> 980
#> 925
#> 872
#> 821
#> 772
#> 725
#> 680
#> 637
#> 596
#> 557
#> 520
#> 485
#> 452
#> 421
#> 392
#> 365
#> 340
#> 317
#> 296
#> 277
#> 260
#> 245
#> 232
#> 221
#> 212
#> 205
#> 200
#> 197
#> 196
#> 197
#> 200
#> 205
#> 212
#> 221
#> 232
#> 245
#> 260
#> 277
#> 296
#> 317
#> 340
#> 365
#> 392
#> 421
#> 452
#> 485
#> 520
#> 557
#> 596
#> 637
#> 680
#> 725
#> 772
#> 821
#> 872
#> 925
#> 980
#> 1037
#> 1096
#> 1157
#> 1220
#> 1249
#> 1186
#> 1125
#> 1066
#> 1009
#> 954
#> 901
#> 850
#> 801
#> 754
#> 709
#> 666
#> 625
#> 586
#> 549
#> 514
#> 481
#> 450
#> 421
#> 394
#> 369
#> 346
#> 325
#> 306
#> 289
#> 274
#> 261
#> 250
#> 241
#> 234
#> 229
#> 226
#> 225
#> 226
#> 229
#> 234
#> 241
#> 250
#> 261
#> 274
#> 289
#> 306
#> 325
#> 346
#> 369
#> 394
#> 421
#> 450
#> 481
#> 514
#> 549
#> 586
#> 625
#> 666
#> 709
#> 754
#> 801
#> 850
#> 901
#> 954
#> 1009
#> 1066
#> 1125
#> 1186
#> 1249
#> 1280
#> 1217
#> 1156
#> 1097
#> 1040
#> 985
#> 932
#> 881
#> 832
#> 785
#> 740
#> 697
#> 656
#> 617
#> 580
#> 545
#> 512
#> 481
#> 452
#> 425
#> 400
#> 377
#> 356
#> 337
#> 320
#> 305
#> 292
#> 281
#> 272
#> 265
#> 260
#> 257
#> 256
#> 257
#> 260
#> 265
#> 272
#> 281
#> 292
#> 305
#> 320
#> 337
#> 356
#> 377
#> 400
#> 425
#> 452
#> 481
#> 512
#> 545
#> 580
#> 617
#> 656
#> 697
#> 740
#> 785
#> 832
#> 881
#> 932
#> 985
#> 1040
#> 1097
#> 1156
#> 1217
#> 1280
#> 1313
#> 1250
#> 1189
#> 1130
#> 1073
#> 1018
#> 965
#> 914
#> 865
#> 818
#> 773
#> 730
#> 689
#> 650
#> 613
#> 578
#> 545
#> 514
#> 485
#> 458
#> 433
#> 410
#> 389
#> 370
#> 353
#> 338
#> 325
#> 314
#> 305
#> 298
#> 293
#> 290
#> 289
#> 290
#> 293
#> 298
#> 305
#> 314
#> 325
#> 338
#> 353
#> 370
#> 389
#> 410
#> 433
#> 458
#> 485
#> 514
#> 545
#> 578
#> 613
#> 650
#> 689
#> 730
#> 773
#> 818
#> 865
#> 914
#> 965
#> 1018
#> 1073
#> 1130
#> 1189
#> 1250
#> 1313
#> 1348
#> 1285
#> 1224
#> 1165
#> 1108
#> 1053
#> 1000
#> 949
#> 900
#> 853
#> 808
#> 765
#> 724
#> 685
#> 648
#> 613
#> 580
#> 549
#> 520
#> 493
#> 468
#> 445
#> 424
#> 405
#> 388
#> 373
#> 360
#> 349
#> 340
#> 333
#> 328
#> 325
#> 324
#> 325
#> 328
#> 333
#> 340
#> 349
#> 360
#> 373
#> 388
#> 405
#> 424
#> 445
#> 468
#> 493
#> 520
#> 549
#> 580
#> 613
#> 648
#> 685
#> 724
#> 765
#> 808
#> 853
#> 900
#> 949
#> 1000
#> 1053
#> 1108
#> 1165
#> 1224
#> 1285
#> 1348
#> 1385
#> 1322
#> 1261
#> 1202
#> 1145
#> 1090
#> 1037
#> 986
#> 937
#> 890
#> 845
#> 802
#> 761
#> 722
#> 685
#> 650
#> 617
#> 586
#> 557
#> 530
#> 505
#> 482
#> 461
#> 442
#> 425
#> 410
#> 397
#> 386
#> 377
#> 370
#> 365
#> 362
#> 361
#> 362
#> 365
#> 370
#> 377
#> 386
#> 397
#> 410
#> 425
#> 442
#> 461
#> 482
#> 505
#> 530
#> 557
#> 586
#> 617
#> 650
#> 685
#> 722
#> 761
#> 802
#> 845
#> 890
#> 937
#> 986
#> 1037
#> 1090
#> 1145
#> 1202
#> 1261
#> 1322
#> 1385
#> 1424
#> 1361
#> 1300
#> 1241
#> 1184
#> 1129
#> 1076
#> 1025
#> 976
#> 929
#> 884
#> 841
#> 800
#> 761
#> 724
#> 689
#> 656
#> 625
#> 596
#> 569
#> 544
#> 521
#> 500
#> 481
#> 464
#> 449
#> 436
#> 425
#> 416
#> 409
#> 404
#> 401
#> 400
#> 401
#> 404
#> 409
#> 416
#> 425
#> 436
#> 449
#> 464
#> 481
#> 500
#> 521
#> 544
#> 569
#> 596
#> 625
#> 656
#> 689
#> 724
#> 761
#> 800
#> 841
#> 884
#> 929
#> 976
#> 1025
#> 1076
#> 1129
#> 1184
#> 1241
#> 1300
#> 1361
#> 1424
#> 1465
#> 1402
#> 1341
#> 1282
#> 1225
#> 1170
#> 1117
#> 1066
#> 1017
#> 970
#> 925
#> 882
#> 841
#> 802
#> 765
#> 730
#> 697
#> 666
#> 637
#> 610
#> 585
#> 562
#> 541
#> 522
#> 505
#> 490
#> 477
#> 466
#> 457
#> 450
#> 445
#> 442
#> 441
#> 442
#> 445
#> 450
#> 457
#> 466
#> 477
#> 490
#> 505
#> 522
#> 541
#> 562
#> 585
#> 610
#> 637
#> 666
#> 697
#> 730
#> 765
#> 802
#> 841
#> 882
#> 925
#> 970
#> 1017
#> 1066
#> 1117
#> 1170
#> 1225
#> 1282
#> 1341
#> 1402
#> 1465
#> 1508
#> 1445
#> 1384
#> 1325
#> 1268
#> 1213
#> 1160
#> 1109
#> 1060
#> 1013
#> 968
#> 925
#> 884
#> 845
#> 808
#> 773
#> 740
#> 709
#> 680
#> 653
#> 628
#> 605
#> 584
#> 565
#> 548
#> 533
#> 520
#> 509
#> 500
#> 493
#> 488
#> 485
#> 484
#> 485
#> 488
#> 493
#> 500
#> 509
#> 520
#> 533
#> 548
#> 565
#> 584
#> 605
#> 628
#> 653
#> 680
#> 709
#> 740
#> 773
#> 808
#> 845
#> 884
#> 925
#> 968
#> 1013
#> 1060
#> 1109
#> 1160
#> 1213
#> 1268
#> 1325
#> 1384
#> 1445
#> 1508
#> 1553
#> 1490
#> 1429
#> 1370
#> 1313
#> 1258
#> 1205
#> 1154
#> 1105
#> 1058
#> 1013
#> 970
#> 929
#> 890
#> 853
#> 818
#> 785
#> 754
#> 725
#> 698
#> 673
#> 650
#> 629
#> 610
#> 593
#> 578
#> 565
#> 554
#> 545
#> 538
#> 533
#> 530
#> 529
#> 530
#> 533
#> 538
#> 545
#> 554
#> 565
#> 578
#> 593
#> 610
#> 629
#> 650
#> 673
#> 698
#> 725
#> 754
#> 785
#> 818
#> 853
#> 890
#> 929
#> 970
#> 1013
#> 1058
#> 1105
#> 1154
#> 1205
#> 1258
#> 1313
#> 1370
#> 1429
#> 1490
#> 1553
#> 1600
#> 1537
#> 1476
#> 1417
#> 1360
#> 1305
#> 1252
#> 1201
#> 1152
#> 1105
#> 1060
#> 1017
#> 976
#> 937
#> 900
#> 865
#> 832
#> 801
#> 772
#> 745
#> 720
#> 697
#> 676
#> 657
#> 640
#> 625
#> 612
#> 601
#> 592
#> 585
#> 580
#> 577
#> 576
#> 577
#> 580
#> 585
#> 592
#> 601
#> 612
#> 625
#> 640
#> 657
#> 676
#> 697
#> 720
#> 745
#> 772
#> 801
#> 832
#> 865
#> 900
#> 937
#> 976
#> 1017
#> 1060
#> 1105
#> 1152
#> 1201
#> 1252
#> 1305
#> 1360
#> 1417
#> 1476
#> 1537
#> 1600
#> 1649
#> 1586
#> 1525
#> 1466
#> 1409
#> 1354
#> 1301
#> 1250
#> 1201
#> 1154
#> 1109
#> 1066
#> 1025
#> 986
#> 949
#> 914
#> 881
#> 850
#> 821
#> 794
#> 769
#> 746
#> 725
#> 706
#> 689
#> 674
#> 661
#> 650
#> 641
#> 634
#> 629
#> 626
#> 625
#> 626
#> 629
#> 634
#> 641
#> 650
#> 661
#> 674
#> 689
#> 706
#> 725
#> 746
#> 769
#> 794
#> 821
#> 850
#> 881
#> 914
#> 949
#> 986
#> 1025
#> 1066
#> 1109
#> 1154
#> 1201
#> 1250
#> 1301
#> 1354
#> 1409
#> 1466
#> 1525
#> 1586
#> 1649
#> 1700
#> 1637
#> 1576
#> 1517
#> 1460
#> 1405
#> 1352
#> 1301
#> 1252
#> 1205
#> 1160
#> 1117
#> 1076
#> 1037
#> 1000
#> 965
#> 932
#> 901
#> 872
#> 845
#> 820
#> 797
#> 776
#> 757
#> 740
#> 725
#> 712
#> 701
#> 692
#> 685
#> 680
#> 677
#> 676
#> 677
#> 680
#> 685
#> 692
#> 701
#> 712
#> 725
#> 740
#> 757
#> 776
#> 797
#> 820
#> 845
#> 872
#> 901
#> 932
#> 965
#> 1000
#> 1037
#> 1076
#> 1117
#> 1160
#> 1205
#> 1252
#> 1301
#> 1352
#> 1405
#> 1460
#> 1517
#> 1576
#> 1637
#> 1700
#> 1753
#> 1690
#> 1629
#> 1570
#> 1513
#> 1458
#> 1405
#> 1354
#> 1305
#> 1258
#> 1213
#> 1170
#> 1129
#> 1090
#> 1053
#> 1018
#> 985
#> 954
#> 925
#> 898
#> 873
#> 850
#> 829
#> 810
#> 793
#> 778
#> 765
#> 754
#> 745
#> 738
#> 733
#> 730
#> 729
#> 730
#> 733
#> 738
#> 745
#> 754
#> 765
#> 778
#> 793
#> 810
#> 829
#> 850
#> 873
#> 898
#> 925
#> 954
#> 985
#> 1018
#> 1053
#> 1090
#> 1129
#> 1170
#> 1213
#> 1258
#> 1305
#> 1354
#> 1405
#> 1458
#> 1513
#> 1570
#> 1629
#> 1690
#> 1753
#> 1808
#> 1745
#> 1684
#> 1625
#> 1568
#> 1513
#> 1460
#> 1409
#> 1360
#> 1313
#> 1268
#> 1225
#> 1184
#> 1145
#> 1108
#> 1073
#> 1040
#> 1009
#> 980
#> 953
#> 928
#> 905
#> 884
#> 865
#> 848
#> 833
#> 820
#> 809
#> 800
#> 793
#> 788
#> 785
#> 784
#> 785
#> 788
#> 793
#> 800
#> 809
#> 820
#> 833
#> 848
#> 865
#> 884
#> 905
#> 928
#> 953
#> 980
#> 1009
#> 1040
#> 1073
#> 1108
#> 1145
#> 1184
#> 1225
#> 1268
#> 1313
#> 1360
#> 1409
#> 1460
#> 1513
#> 1568
#> 1625
#> 1684
#> 1745
#> 1808
#> 1865
#> 1802
#> 1741
#> 1682
#> 1625
#> 1570
#> 1517
#> 1466
#> 1417
#> 1370
#> 1325
#> 1282
#> 1241
#> 1202
#> 1165
#> 1130
#> 1097
#> 1066
#> 1037
#> 1010
#> 985
#> 962
#> 941
#> 922
#> 905
#> 890
#> 877
#> 866
#> 857
#> 850
#> 845
#> 842
#> 841
#> 842
#> 845
#> 850
#> 857
#> 866
#> 877
#> 890
#> 905
#> 922
#> 941
#> 962
#> 985
#> 1010
#> 1037
#> 1066
#> 1097
#> 1130
#> 1165
#> 1202
#> 1241
#> 1282
#> 1325
#> 1370
#> 1417
#> 1466
#> 1517
#> 1570
#> 1625
#> 1682
#> 1741
#> 1802
#> 1865
#> 1924
#> 1861
#> 1800
#> 1741
#> 1684
#> 1629
#> 1576
#> 1525
#> 1476
#> 1429
#> 1384
#> 1341
#> 1300
#> 1261
#> 1224
#> 1189
#> 1156
#> 1125
#> 1096
#> 1069
#> 1044
#> 1021
#> 1000
#> 981
#> 964
#> 949
#> 936
#> 925
#> 916
#> 909
#> 904
#> 901
#> 900
#> 901
#> 904
#> 909
#> 916
#> 925
#> 936
#> 949
#> 964
#> 981
#> 1000
#> 1021
#> 1044
#> 1069
#> 1096
#> 1125
#> 1156
#> 1189
#> 1224
#> 1261
#> 1300
#> 1341
#> 1384
#> 1429
#> 1476
#> 1525
#> 1576
#> 1629
#> 1684
#> 1741
#> 1800
#> 1861
#> 1924
#> 1985
#> 1922
#> 1861
#> 1802
#> 1745
#> 1690
#> 1637
#> 1586
#> 1537
#> 1490
#> 1445
#> 1402
#> 1361
#> 1322
#> 1285
#> 1250
#> 1217
#> 1186
#> 1157
#> 1130
#> 1105
#> 1082
#> 1061
#> 1042
#> 1025
#> 1010
#> 997
#> 986
#> 977
#> 970
#> 965
#> 962
#> 961
#> 962
#> 965
#> 970
#> 977
#> 986
#> 997
#> 1010
#> 1025
#> 1042
#> 1061
#> 1082
#> 1105
#> 1130
#> 1157
#> 1186
#> 1217
#> 1250
#> 1285
#> 1322
#> 1361
#> 1402
#> 1445
#> 1490
#> 1537
#> 1586
#> 1637
#> 1690
#> 1745
#> 1802
#> 1861
#> 1922
#> 1985
#> 2048
#> 1985
#> 1924
#> 1865
#> 1808
#> 1753
#> 1700
#> 1649
#> 1600
#> 1553
#> 1508
#> 1465
#> 1424
#> 1385
#> 1348
#> 1313
#> 1280
#> 1249
#> 1220
#> 1193
#> 1168
#> 1145
#> 1124
#> 1105
#> 1088
#> 1073
#> 1060
#> 1049
#> 1040
#> 1033
#> 1028
#> 1025
#> 1024
#> 1025
#> 1028
#> 1033
#> 1040
#> 1049
#> 1060
#> 1073
#> 1088
#> 1105
#> 1124
#> 1145
#> 1168
#> 1193
#> 1220
#> 1249
#> 1280
#> 1313
#> 1348
#> 1385
#> 1424
#> 1465
#> 1508
#> 1553
#> 1600
#> 1649
#> 1700
#> 1753
#> 1808
#> 1865
#> 1924
#> 1985
#> 2048
#> [[1]]
#> Marked planar point pattern: 180 points
#> Multitype, with levels = 1, 2 
#> window: rectangle = [0, 1] x [0, 1] units
#> 

## ------------------------------------------------
## Method `TwoMarkDPP$print`
## ------------------------------------------------

mod <- TwoMarkDPP$new()
mod
#> 
#> ── Two-mark planar Gauss determinantal point process ───────────────────────────
#> • Intensity of the 1st marginal DPP: ?
#> • Intensity of the 2nd marginal DPP: ?
#> • Repulsion rate of the 1st marginal DPP: ?
#> • Repulsion rate of the 2nd marginal DPP: ?
#> • Repulsion rate between marks: ?
#> • Correlation between marks: ?
#> • Window size: ?