Estimation of Two-Mark Planar DPPs via the Pair Correlation Function

Estimation of Two-Mark Planar DPPs via the Pair Correlation Function

Usage

fit_via_pcf(
  X,
  model = "Gauss",
  rmin_alpha = 2,
  rmin_alpha12 = 2,
  rmin_tau = 2,
  q = 1,
  p = 2,
  divisor_marginal = "d",
  divisor_cross = "d",
  method = "profiling",
  B = 0L,
  full_bootstrap = TRUE,
  params = NULL
)

Arguments

X

An object of class spatstat.geom::ppp specifying a planar point pattern.

model

A string specifying the model to be fitted. Choices are "Gauss" or "Bessel". Defaults to "Gauss".

rmin_alpha

The lower bound on distances that should be taken into account for estimating marginal alpha parameters (default: index 2).

rmin_alpha12

The lower bound on distances that should be taken into account for estimating the crossing alpha parameter (default: index 2).

rmin_tau

The lower bound on distances that should be taken into account for estimating the correlation (default: index 2).

q

Power for pointwise evaluations of the PCF. Defaults to 0.5.

p

Power for the distance between empirical and moodel-based PCF values. Defaults to 2.

divisor_marginal

Choice of divisor in the estimation formula. Choices are "r" or "d". See Section Empirical estimation of the pair correlation function for more details. Defaults to "d".

divisor_cross

Choice of divisor in the estimation formula. Choices are "r" or "d". See Section Empirical estimation of the pair correlation function for more details. Defaults to "d".

method

A character string specifying the estimation method between "profiling" and "direct". Defaults to "profiling".

B

An integer value specifying the number of samples to be generated in the bootstrap procedure to approximate the distribution of the test statistics when testing for absence of correlation between marks. Defaults to 0L which does not perform the test at all.

full_bootstrap

A boolean specifying whether marginal parameters should be re-estimated when computing the bootstrapped distribution of the tau statistic. Defaults to TRUE.

params

A length-4 numeric vector specifying values for the marginal parameters if known. The order needs to be rho1, rho2, alpha1 and alpha2. Defaults to NULL, in which case, they are estimated.

Value

A list with the estimated model parameters in the following order: rho1, rho2, alpha1, alpha2, k12, alpha12 and tau. Additional information pertaining to the test for absence of correlation between marks are returned in the list as well.

Empirical estimation of the pair correlation function

The empirical PCF is computed as a kernel estimate of the PCF in which the contribution from an interpoint distance \(d_{ij}\) to the estimate of \(g(r)\) is divided:

either by \(r\) using optional argument divisor = "r" in the functions spatstat.explore::pcf() and spatstat.explore::pcfcross();

or by \(d_{ij}\) using optional argument divisor = "d" in the functions spatstat.explore::pcf() and spatstat.explore::pcfcross(); it is intended to improve the bias of the estimator when \(r\) is close to zero.

Examples

fit_via_pcf(sim_gauss0[[1]])
#> $rho1
#> [1] 100
#> 
#> $rho2
#> [1] 86
#> 
#> $alpha1
#> [1] 0.0323542
#> 
#> $alpha2
#> [1] 0.02833893
#> 
#> $k12
#> [1] 0
#> 
#> $tau
#> [1] 0
#> 
#> $alpha12
#> [1] 0.0304129
#> 
#> $fmin
#> [1] 0.0008385533
#> 
#> $stat_np_obs
#> [1] 0.0008385533
#> 
#> $stat_p_obs
#> [1] 0
#> 
#> $stat_np_boots
#> NULL
#> 
#> $stat_p_boots
#> NULL
#> 
#> $pvalue_np
#> NULL
#> 
#> $pvalue_p
#> NULL
#>