Estimation of multi-mark DPP via Maximum Likelihood

Estimation of multi-mark DPP via Maximum Likelihood

Usage

mle(
  data,
  initial_guess = NULL,
  fixed_marginal_parameters = FALSE,
  model = "gauss",
  optimizer = "bobyqa",
  num_threads = 1,
  N = 512,
  verbose_level = 0
)

Arguments

data

An object of class spatstat.geom::ppp specifying an observed planar determinantal point process.

initial_guess

A numeric vector specifying an initial guess for the model parameters that maximize the likelihood. Defaults to NULL, which initializes at \(0.5\) all parameters after suitable transformation into \([0, 1]\). If provided, expected order is c(alpha1, alpha2, alpha12, tau).

fixed_marginal_parameters

A boolean value specifying whether the marginal parameters should be estimated separately using the marginal likelihood and then fixed to further estimate the cross parameters. Defaults to FALSE.

model

A string specifying a DPP model. Choices are "gauss" or "bessel". Defaults to "gauss".

optimizer

A string specifying one of the available derivative-free optimizers in NLOpt. Defaults to "bobyqa".

num_threads

An integer value specifying the number of thread to run on. Defaults to 1L.

N

An integer value specifying the maximum truncation index for Fourier transform. Defaults to 512L.

verbose_level

An integer value specifying the display information during optimization. Choose 0L for no information, 1L for global information at likelihood setup or 2L for detailed information at each function evaluation. Defaults to `0L.

Value

A list with two components:

par (optimized model parameters);

value (-2 logLik at the maximum likelihood).

Examples

opt <- mle(sim_gauss5[[1]])