Access and assignment methods for muscadet objects

Simplified access to omic datasets and slots inside muscadet objects.

Assign new data in a muscadet or muscomic object. For muscadet objects, the omic datasets in the omics slot can be directly reassigned.

Usage

# S3 method for class 'muscadet'
x[i, ...]

# S3 method for class 'muscomic'
x[i, ...]

# S3 method for class 'muscadet'
x$name

# S3 method for class 'muscomic'
x$name

# S3 method for class 'muscadet'
x[i] <- value

# S3 method for class 'muscomic'
x[i] <- value

# S3 method for class 'muscadet'
x$i <- value

# S3 method for class 'muscomic'
x$i <- value

Arguments

x

A muscadet or muscomic object.

i

The name of the slot (or omic).

...

Other arguments (ignored).

name

The name of the slot (or omic).

value

The new value to assign.

Value

The selected slot or the omic dataset (muscomic object) for muscadet objects. The selected slot for muscomic objects.

The updated muscadet or muscomic object.

Examples

# Load example muscadet object
data(muscadet_obj)
# Access to muscadet omics or slots
muscadet_obj["ATAC"]
#> A muscomic object of type ATAC labelled scATAC-seq containing: 
#>  log.ratio coverage data matrix 
#>  112 cells 
#>  133 features: windows of peaks 
#>  691 variant positions 
muscadet_obj["genome"]
#> [1] "hg38"

# Load example muscadet object
data(muscadet_obj)
# Access to muscomic slots
muscadet_obj["ATAC"]["label.omic"]
#> [1] "scATAC-seq"

# Load example muscadet object
data(muscadet_obj)
# Access to muscadet omics or slots
muscadet_obj$ATAC
#> A muscomic object of type ATAC labelled scATAC-seq containing: 
#>  log.ratio coverage data matrix 
#>  112 cells 
#>  133 features: windows of peaks 
#>  691 variant positions 
muscadet_obj$genome
#> [1] "hg38"

# Load example muscadet object
data(muscadet_obj)
# Access to muscomic slots
muscadet_obj$ATAC$label.omic
#> [1] "scATAC-seq"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscadet object
muscadet_obj["genome"] <- "hg38"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscomic object
muscadet_obj["ATAC"]["label.omic"] <- "scATAC-seq"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscadet object
muscadet_obj$genome <- "hg38"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscomic object
muscadet_obj$ATAC$label.omic <- "scATAC-seq"